Script - Adding play when click in button
Today I added the feature to show the video when the button is clicked. To do this, I set some properties in video html tag to change this when the click event is trigged. The visibility property was one of these property that I used to make the video appear. Now, I have to fix somes settings in animations syntax to make the video appear smothly, some settings in widht and height to user see controls and make something to close the video. I'm thinking to add some div and caught this with JS or some input with his pseudo-elements and make this only in the CSS.